+++ libsel4/sel4_plat_include/hobgoblin/sel4/plat/api/constants.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright 2024, Codasib GmbH
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#pragma once
+
+#include <sel4/config.h>
+
+/* nothing here */